+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Add a compact facing-direction indicator for patrol NPCs. (#48)
 - Dev: Switch to Nerixyz/actionsx-prettier. (#49)
 
 ## 1.6